reverting optimizer tests
